requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cycler==0.10.0
importlib-metadata==0.23
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==7.2.0
numpy==1.17.2
packaging==19.2
Pillow==6.1.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.3
python-dateutil==2.8.0
six==1.12.0
wcwidth==0.1.7
zipp==0.6.0
